Many different changes like transition to "click" lib from clear python inputs, deleted main function and close
